Separated adventure drawing creation funcs and interactions and style definitions.
